requirements.txt
python-telegram-bot
supabase
google-genai
python-dotenv
redis
